+++ bsps/shared/dev/serial/ns16550.c
@@ -850,7 +850,7 @@ ssize_t ns16550_write_support_polled(
   size_t      len
 )
 {
-  int nwrite = 0;
+  size_t nwrite = 0;
 
   /*
    * poll each byte in the string out of the port.
